+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OzoneConfigUtil.java
@@ -19,20 +19,10 @@ package org.apache.hadoop.ozone.om;
 
 import org.apache.hadoop.hdds.client.DefaultReplicationConfig;
 import org.apache.hadoop.hdds.client.ReplicationConfig;
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.ozone.om.exceptions.OMException;
-import org.apache.hadoop.security.UserGroupInformation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.util.Collection;
-
-import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_ADMINISTRATORS;
-import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_ADMINISTRATORS_GROUPS;
-import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_S3_ADMINISTRATORS;
-import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_S3_ADMINISTRATORS_GROUPS;
 
 /**
  * Utility class for ozone configurations.
@@ -41,38 +31,6 @@ public final class OzoneConfigUtil {
   static final Logger LOG =
       LoggerFactory.getLogger(OzoneConfigUtil.class);
   private OzoneConfigUtil() {
-  }
-
-  /**
-   * Return list of s3 administrators prop from config.
-   *
-   * If ozone.s3.administrators value is empty string or unset,
-   * defaults to ozone.administrators value.
-   */
-  public static Collection<String> getS3AdminsFromConfig(OzoneConfiguration conf)
-          throws IOException {
-    Collection<String> ozAdmins =
-            conf.getTrimmedStringCollection(OZONE_S3_ADMINISTRATORS);
-    if (ozAdmins == null || ozAdmins.isEmpty()) {
-      ozAdmins = conf.getTrimmedStringCollection(OZONE_ADMINISTRATORS);
-    }
-    String omSPN = UserGroupInformation.getCurrentUser().getShortUserName();
-    if (!ozAdmins.contains(omSPN)) {
-      ozAdmins.add(omSPN);
-    }
-    return ozAdmins;
-  }
-
-  public static Collection<String> getS3AdminsGroupsFromConfig(
-      OzoneConfiguration conf) {
-    Collection<String> s3AdminsGroup =
-            conf.getTrimmedStringCollection(OZONE_S3_ADMINISTRATORS_GROUPS);
-    if (s3AdminsGroup.isEmpty() && conf
-        .getTrimmedStringCollection(OZONE_S3_ADMINISTRATORS).isEmpty()) {
-      s3AdminsGroup = conf
-              .getTrimmedStringCollection(OZONE_ADMINISTRATORS_GROUPS);
-    }
-    return s3AdminsGroup;
   }
 
   public static ReplicationConfig resolveReplicationConfigPreference(
