+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/TestOzoneConfigUtil.java
@@ -29,7 +29,6 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import java.util.Arrays;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -118,43 +117,5 @@ public class TestOzoneConfigUtil {
     // Client has no preference of type and bucket has ratis defaults, so it
     // should return ratis.
     assertEquals(ratisReplicationConfig, replicationConfig);
-  }
-
-  @Test
-  public void testS3AdminExtraction() throws IOException {
-    OzoneConfiguration configuration = new OzoneConfiguration();
-    configuration.set(OzoneConfigKeys.OZONE_S3_ADMINISTRATORS, "alice,bob");
-
-    assertThat(OzoneConfigUtil.getS3AdminsFromConfig(configuration))
-        .containsAll(Arrays.asList("alice", "bob"));
-  }
-
-  @Test
-  public void testS3AdminExtractionWithFallback() throws IOException {
-    OzoneConfiguration configuration = new OzoneConfiguration();
-    configuration.set(OzoneConfigKeys.OZONE_ADMINISTRATORS, "alice,bob");
-
-    assertThat(OzoneConfigUtil.getS3AdminsFromConfig(configuration))
-        .containsAll(Arrays.asList("alice", "bob"));
-  }
-
-  @Test
-  public void testS3AdminGroupExtraction() {
-    OzoneConfiguration configuration = new OzoneConfiguration();
-    configuration.set(OzoneConfigKeys.OZONE_S3_ADMINISTRATORS_GROUPS,
-        "test1, test2");
-
-    assertThat(OzoneConfigUtil.getS3AdminsGroupsFromConfig(configuration))
-        .containsAll(Arrays.asList("test1", "test2"));
-  }
-
-  @Test
-  public void testS3AdminGroupExtractionWithFallback() {
-    OzoneConfiguration configuration = new OzoneConfiguration();
-    configuration.set(OzoneConfigKeys.OZONE_ADMINISTRATORS_GROUPS,
-        "test1, test2");
-
-    assertThat(OzoneConfigUtil.getS3AdminsGroupsFromConfig(configuration))
-        .containsAll(Arrays.asList("test1", "test2"));
   }
 }
